+++ pyreadstat/pyreadstat.pyx
@@ -33,7 +33,7 @@ from copy import deepcopy
 # Parsing functions
 
 def read_sas7bdat(str filename_path, metadataonly=False, dates_as_pandas_datetime=False, catalog_file=None,
-                  formats_as_category=True, str encoding=None, list usecols=None):
+                  formats_as_category=True, str encoding=None, list usecols=None, user_missing=False):
     r"""
     Read a SAS sas7bdat file.
     It accepts the path to a sas7bcat.
@@ -60,6 +60,11 @@ def read_sas7bdat(str filename_path, metadataonly=False, dates_as_pandas_datetim
             iconv-compatible name
         usecols : list, optional
             a list with column names to read from the file. Only those columns will be imported. Case sensitive!
+        user_missing : bool, optional
+            by default False, in this case user defined missing values are delivered as nan. If true, the missing values
+            will be deliver as is, and an extra piece of information will be set in the metadata (missing_user_values)
+            to be able to interpret those values as missing. At the time of writing this, it does not work for all files
+            (maybe only for those produced on windows, but not on unix)
 
     Returns
     -------
@@ -80,6 +85,8 @@ def read_sas7bdat(str filename_path, metadataonly=False, dates_as_pandas_datetim
         dates_as_pandas = 1
 
     cdef bint usernan = 0
+    if user_missing:
+        usernan = 1
     
     cdef py_file_format file_format = _readstat_parser.FILE_FORMAT_SAS
     data_frame, metadata = run_conversion(filename_path, file_format, readstat_parse_sas7bdat, encoding, metaonly,